refactor: Change `expense` type from `int` to `double` in `ExpenseInfo`
This commit updates the `expense` field in the `ExpenseInfo` model to support decimal values by changing its type from `int` to `double`. This includes updating the JSON serialization logic and the associated view model to ensure consistent handling of floating-point numbers.

Specific changes include:

*   **ExpenseInfo Model**:
    *   Changed the type of the `expense` field from `int` to `double`.
    *   Added `_expenseFromJson` and `_expenseToJson` helper functions to safely handle numeric conversions during serialization (supporting both `int` and `double` inputs from JSON).
    *   Regenerated `expense_info.g.dart` and `expense_info.freezed.dart` to reflect the type change and custom JSON keys.
*   **AddEditExpenseViewModel**: Updated validation logic to use `double.tryParse` instead of `int.tryParse` when processing expense input.

+++ lib/ui/main/expenses/add_edit/view_models/add_edit_expense_viewmodel.dart
@@ -129,7 +129,7 @@ class AddEditExpenseViewModel extends ChangeNotifier {
     }
 
     /* 金額をチェックする */
-    final expenseVal = int.tryParse(expense);
+    final expenseVal = double.tryParse(expense);
     if (expenseVal == null) {
       print("expenseが数値に変換できない");
       return ResultInfo.failed(error: ErrorInfo(errorMessage: "金額が入力が不適切です"));
